Update object methods to pass bucket directly to s3HTTP
Currently all of the object methods have hard-wired the
convention used to indicate the bucket in the url.  Further,
it appears in my test at least that this subdomain convention
fails with buckets that have `.` in the bucket name.  To address
this, I have remapped each of the object methods not to use the
`url` parameter of s3HTTP (which I think should be removed),
but instead to just pass the bucket name to the s3HTTP method.

Even if we do decide to stick with the subdomain convention, it
should be set in one place in the s3HTTP method so it is easy to
switch from one to the other, and not be hardwired into each of
the object methods.